pyproject.toml
[project]
name = "scorer-msme-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "lightrag[api]>=0.0.0a2",
    "numpy>=2.3.2",
    "orjson>=3.11.1",
    "pdfplumber>=0.11.7",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.1.0",
    "uvicorn[standard]>=0.35.0",
    "aiohttp>=3.9.0",
]
